delete form service, using just validator

+++ blog-server/src/App/BlogBundle/Service/Impl/FileUploaderImpl.php
@@ -23,11 +23,10 @@ class FileUploaderImpl implements FileUploader
 
     public function delete($fileName)
     {
-        if (file_exists($this->getTargetDir() . $fileName)) {
+        if (!empty($fileName) && file_exists($this->getTargetDir() . $fileName)) {
             unlink($this->getTargetDir() . $fileName);
         }
     }
-
 
     public function getTargetDir()
     {
